rename files and add navio target
