+++ src/drivers/navio_sysfs_pwm_out/navio_sysfs_pwm_out.cpp
@@ -50,7 +50,7 @@
 #include <systemlib/param/param.h>
 #include <systemlib/pwm_limit/pwm_limit.h>
 
-namespace pwm_out
+namespace navio_sysfs_pwm_out
 {
 static px4_task_t _task_handle = -1;
 volatile bool _task_should_exit = false;
@@ -431,12 +431,12 @@ void usage()
 	PX4_INFO("       pwm_out status");
 }
 
-} // namespace pwm_out
+} // namespace navio_sysfs_pwm_out
 
 /* driver 'main' command */
-extern "C" __EXPORT int pwm_out_main(int argc, char *argv[]);
+extern "C" __EXPORT int navio_sysfs_pwm_out_main(int argc, char *argv[]);
 
-int pwm_out_main(int argc, char *argv[])
+int navio_sysfs_pwm_out_main(int argc, char *argv[])
 {
 	const char *device = nullptr;
 	int ch;
@@ -456,43 +456,43 @@ int pwm_out_main(int argc, char *argv[])
 		switch (ch) {
 		case 'd':
 			device = myoptarg;
-			strncpy(pwm_out::_device, device, strlen(device));
+			strncpy(navio_sysfs_pwm_out::_device, device, strlen(device));
 			break;
 		}
 	}
 
 	// gets the parameters for the esc's pwm
-	param_get(param_find("PWM_DISARMED"), &pwm_out::_pwm_disarmed);
-	param_get(param_find("PWM_MIN"), &pwm_out::_pwm_min);
-	param_get(param_find("PWM_MAX"), &pwm_out::_pwm_max);
+	param_get(param_find("PWM_DISARMED"), &navio_sysfs_pwm_out::_pwm_disarmed);
+	param_get(param_find("PWM_MIN"), &navio_sysfs_pwm_out::_pwm_min);
+	param_get(param_find("PWM_MAX"), &navio_sysfs_pwm_out::_pwm_max);
 
 	/*
 	 * Start/load the driver.
 	 */
 	if (!strcmp(verb, "start")) {
-		if (pwm_out::_is_running) {
+		if (navio_sysfs_pwm_out::_is_running) {
 			PX4_WARN("pwm_out already running");
 			return 1;
 		}
 
-		pwm_out::start();
+		navio_sysfs_pwm_out::start();
 	}
 
 	else if (!strcmp(verb, "stop")) {
-		if (!pwm_out::_is_running) {
+		if (!navio_sysfs_pwm_out::_is_running) {
 			PX4_WARN("pwm_out is not running");
 			return 1;
 		}
 
-		pwm_out::stop();
+		navio_sysfs_pwm_out::stop();
 	}
 
 	else if (!strcmp(verb, "status")) {
-		PX4_WARN("pwm_out is %s", pwm_out::_is_running ? "running" : "not running");
+		PX4_WARN("pwm_out is %s", navio_sysfs_pwm_out::_is_running ? "running" : "not running");
 		return 0;
 
 	} else {
-		pwm_out::usage();
+		navio_sysfs_pwm_out::usage();
 		return 1;
 	}
